requirements.txt
google-api-core==1.4.1
google-auth==1.5.1
google-cloud-speech==0.36.0
googleapis-common-protos==1.5.3
grpcio==1.15.0
idna==2.7
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
PyAudio==0.2.11
pydub==0.23.0
pytz==2018.5
requests==2.19.1
rsa==4.0
six==1.11.0
SpeechRecognition==3.8.1
urllib3==1.23